# Adding quadratic objective support to NAG solver wrapper