## Description
We have modified the wrapper for the NAG solver to allow quadratic objective information to be passed directly to the solver (without cone transformation). This was done to improve the performance.

## Type of change
- [x] New feature (backwards compatible)
- [ ] New feature (breaking API changes)
- [ ] Bug fix
- [ ] Other (Documentation, CI, ...)

## [Contribution checklist](https://www.cvxpy.org/contributing/index.html#contribution-checklist)
- [x] Add our license to new files. [no new files added]
- [x] Check that your code adheres to our coding style.
- [x] Write unittests.
- [x] Run the unittests and check that they’re passing.
- [x] Run the benchmarks to make sure your change doesn’t introduce a regression.

Benchmark without changes:
================================================= test session starts =================================================
platform win32 -- Python 3.10.8, pytest-7.4.4, pluggy-1.4.0
rootdir: C:\Users\charlotte.silvester\nag\cvxpy
configfile: pyproject.toml
collected 12 items

cvxpy\tests\test_benchmarks.py ss-------------------------------------------------------------------------------
                                  Compilation
-------------------------------------------------------------------------------
(CVXPY) Jan 31 12:52:42 PM: Compiling problem (target solver=ECOS).
(CVXPY) Jan 31 12:52:42 PM: Reduction chain: Dcp2Cone -> CvxAttr2Constr -> ConeMatrixStuffing -> ECOS
(CVXPY) Jan 31 12:52:42 PM: Applying reduction Dcp2Cone
(CVXPY) Jan 31 12:52:42 PM: Applying reduction CvxAttr2Constr
(CVXPY) Jan 31 12:52:42 PM: Applying reduction ConeMatrixStuffing
(CVXPY) Jan 31 12:52:42 PM: Applying reduction ECOS
(CVXPY) Jan 31 12:52:42 PM: Finished problem compilation (took 3.642e-01 seconds).
Issue #1668 regression test
Compilation time:  0.3642103672027588
.least_squares: avg=2.084e-03 s , std=0.000e+00 s (1 iterations)
.ssqp: avg=1.010e-02 s , std=0.000e+00 s (1 iterations)
.small_cone_matrix_stuffing: avg=2.224e-02 s , std=9.971e-03 s (10 iterations)
.small_lp: avg=8.214e-03 s , std=0.000e+00 s (1 iterations)
small_lp_second_time: avg=0.000e+00 s , std=0.000e+00 s (1 iterations)
.sss

============================================ 5 passed, 7 skipped in 1.27s =============================================


Benchmark with changes:

================================================= test session starts =================================================
platform win32 -- Python 3.10.8, pytest-7.4.4, pluggy-1.4.0
rootdir: C:\Users\charlotte.silvester\nag\cvxpy
configfile: pyproject.toml
collected 12 items

cvxpy\tests\test_benchmarks.py ss-------------------------------------------------------------------------------
                                  Compilation
-------------------------------------------------------------------------------
(CVXPY) Jan 31 12:53:36 PM: Compiling problem (target solver=ECOS).
(CVXPY) Jan 31 12:53:36 PM: Reduction chain: Dcp2Cone -> CvxAttr2Constr -> ConeMatrixStuffing -> ECOS
(CVXPY) Jan 31 12:53:36 PM: Applying reduction Dcp2Cone
(CVXPY) Jan 31 12:53:36 PM: Applying reduction CvxAttr2Constr
(CVXPY) Jan 31 12:53:36 PM: Applying reduction ConeMatrixStuffing
(CVXPY) Jan 31 12:53:36 PM: Applying reduction ECOS
(CVXPY) Jan 31 12:53:36 PM: Finished problem compilation (took 3.687e-01 seconds).
Issue #1668 regression test
Compilation time:  0.3687434196472168
.least_squares: avg=0.000e+00 s , std=0.000e+00 s (1 iterations)
.ssqp: avg=1.007e-02 s , std=0.000e+00 s (1 iterations)
.small_cone_matrix_stuffing: avg=2.236e-02 s , std=1.095e-02 s (10 iterations)
.small_lp: avg=1.048e-02 s , std=0.000e+00 s (1 iterations)
small_lp_second_time: avg=0.000e+00 s , std=0.000e+00 s (1 iterations)
.sss

============================================ 5 passed, 7 skipped in 1.25s =============================================