@chasilvester 2024-02-01
I've added a test in my latest commit and all seems to run okay if NAG is not installed. Please let me know if I need to change anything else!